pyproject.toml
[project]
name = "meeting-ai"
version = "0.2.0"
description = "Real-time meeting transcription and AI assistant"
requires-python = ">=3.11"
dependencies = [
    "sounddevice>=0.4",
    "numpy>=1.24",
    "deepgram-sdk>=3.0",
    "python-dotenv>=1.0",
    "twilio>=9.0",
    "anthropic>=0.40",
    "pyngrok>=7.0",
    "websockets>=12.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "anyio>=4.0",
    "pytest-anyio>=0.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
